// src/app/model/city.model.ts
export interface City {
  Taipei: string;
  NewTaipei: string;
  Taoyuan: string;
  Taichung: string;
  Tainan: string;
  Kaohsiung: string;
  Keelung: string;
  Hsinchu: string;
  HsinchuCounty: string;
  MiaoliCounty: string;
  ChanghuaCounty: string;
  NantouCounty: string;
  YunlinCounty: string;
  ChiayiCounty: string;
  Chiayi: string;
  PingtungCounty: string;
  YilanCounty: string;
  HualienCounty: string;
  TaitungCounty: string;
  KinmenCounty: string;
  PenghuCounty: string;
  LienchiangCounty: string;
}